refactor: move functions that update quantity, labels and annotations in a specific file
