+++ controllers/cluster_create.go
@@ -128,7 +128,7 @@ func (r *ClusterReconciler) reconcilePodDisruptionBudget(ctx context.Context, cl
 		)
 	}
 
-	// Reconcile the primary PDB
+	// UpdateQuantity the primary PDB
 	err := r.createOrPatchOwnedPodDisruptionBudget(ctx,
 		cluster,
 		specs.BuildPrimaryPodDisruptionBudget(cluster),
