+++ controllers/replicas.go
@@ -28,7 +28,6 @@ import (
 	"github.com/cloudnative-pg/cloudnative-pg/internal/configuration"
 	"github.com/cloudnative-pg/cloudnative-pg/pkg/management/log"
 	"github.com/cloudnative-pg/cloudnative-pg/pkg/postgres"
-	pvcReconciler "github.com/cloudnative-pg/cloudnative-pg/pkg/reconciler/pvc"
 	"github.com/cloudnative-pg/cloudnative-pg/pkg/specs"
 	"github.com/cloudnative-pg/cloudnative-pg/pkg/utils"
 )
@@ -412,90 +411,6 @@ func (r *ClusterReconciler) updateClusterLabelsOnPods(
 	return nil
 }
 
-// updateClusterAnnotationsOnPVCs we check if we need to add or modify existing annotations specified in the cluster but
-// not existing in the PVCs. We do not support the case of removed annotations from the cluster resource.
-func (r *ClusterReconciler) updateClusterAnnotationsOnPVCs(
-	ctx context.Context,
-	cluster *apiv1.Cluster,
-	pvcs corev1.PersistentVolumeClaimList,
-) error {
-	contextLogger := log.FromContext(ctx)
-
-	for i := range pvcs.Items {
-		pvc := &pvcs.Items[i]
-
-		// if all the required annotations are already set and with the correct value,
-		// we proceed to the next item
-		if utils.IsAnnotationSubset(pvc.Annotations,
-			cluster.Annotations,
-			cluster.GetFixedInheritedLabels(),
-			configuration.Current) &&
-			utils.IsAnnotationAppArmorPresentInObject(&pvc.ObjectMeta, cluster.Annotations) {
-			contextLogger.Debug(
-				"Skipping cluster annotations reconciliation, because they are already present on pvc",
-				"pvc", pvc.Name,
-				"pvcAnnotations", pvc.Annotations,
-				"clusterAnnotations", cluster.Annotations,
-			)
-			continue
-		}
-
-		// otherwise, we add the modified/new annotations to the pvc
-		patch := client.MergeFrom(pvc.DeepCopy())
-		utils.InheritAnnotations(&pvc.ObjectMeta, cluster.Annotations,
-			cluster.GetFixedInheritedAnnotations(), configuration.Current)
-
-		contextLogger.Info("Updating cluster annotations on pvc", "pvc", pvc.Name)
-		if err := r.Patch(ctx, pvc, patch); err != nil {
-			return err
-		}
-		continue
-	}
-
-	return nil
-}
-
-// updateClusterAnnotationsOnPVCs we check if we need to add or modify existing labels specified in the cluster but
-// not existing in the PVCs. We do not support the case of removed labels from the cluster resource.
-func (r *ClusterReconciler) updateClusterLabelsOnPVCs(
-	ctx context.Context,
-	cluster *apiv1.Cluster,
-	pvcs corev1.PersistentVolumeClaimList,
-) error {
-	contextLogger := log.FromContext(ctx)
-
-	for i := range pvcs.Items {
-		pvc := &pvcs.Items[i]
-
-		// if all the required labels are already set and with the correct value,
-		// we proceed to the next item
-		if utils.IsLabelSubset(pvc.Labels,
-			cluster.Labels,
-			cluster.GetFixedInheritedAnnotations(),
-			configuration.Current) {
-			contextLogger.Debug(
-				"Skipping cluster label reconciliation, because they are already present on pvc",
-				"pvc", pvc.Name,
-				"pvcLabels", pvc.Labels,
-				"clusterLabels", cluster.Labels,
-			)
-			continue
-		}
-
-		// otherwise, we add the modified/new labels to the pvc
-		patch := client.MergeFrom(pvc.DeepCopy())
-		utils.InheritLabels(&pvc.ObjectMeta, cluster.Labels, cluster.GetFixedInheritedLabels(), configuration.Current)
-
-		contextLogger.Debug("Updating cluster labels on pvc", "pvc", pvc.Name)
-		if err := r.Patch(ctx, pvc, patch); err != nil {
-			return err
-		}
-		contextLogger.Info("Updated cluster label on pvc", "pvc", pvc.Name)
-	}
-
-	return nil
-}
-
 // Make sure that only the currentPrimary has the label forward write traffic to him
 func (r *ClusterReconciler) updateRoleLabelsOnPods(
 	ctx context.Context,
@@ -581,45 +496,6 @@ func (r *ClusterReconciler) updateOperatorLabelsOnInstances(
 
 		if err := r.Client.Patch(ctx, instance, client.MergeFrom(origInstance)); err != nil {
 			return err
-		}
-	}
-
-	return nil
-}
-
-// updateOperatorLabelsOnPVC ensures that the PVCs have the correct labels
-func (r *ClusterReconciler) updateOperatorLabelsOnPVC(
-	ctx context.Context,
-	instances corev1.PodList,
-	pvcs corev1.PersistentVolumeClaimList,
-) error {
-	for _, pod := range instances.Items {
-		podRole, podHasRole := pod.ObjectMeta.Labels[specs.ClusterRoleLabelName]
-
-		instancePVCs := pvcReconciler.FilterInstancePVCs(pvcs.Items, pod.Spec)
-		for i := range instancePVCs {
-			pvc := &instancePVCs[i]
-			var modified bool
-			// this is needed, because on older versions pvc.labels could be nil
-			if pvc.Labels == nil {
-				pvc.Labels = map[string]string{}
-			}
-
-			origPvc := pvc.DeepCopy()
-			if podHasRole && pvc.ObjectMeta.Labels[specs.ClusterRoleLabelName] != podRole {
-				pvc.Labels[specs.ClusterRoleLabelName] = podRole
-				modified = true
-			}
-			if pvc.ObjectMeta.Labels[utils.InstanceNameLabelName] != pod.Name {
-				pvc.ObjectMeta.Labels[utils.InstanceNameLabelName] = pod.Name
-				modified = true
-			}
-			if !modified {
-				continue
-			}
-			if err := r.Client.Patch(ctx, pvc, client.MergeFrom(origPvc)); err != nil {
-				return err
-			}
 		}
 	}
